requirements2.txt
asttokens==2.4.1
backoff==2.2.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coinbase-advanced-py==1.6.1
colorama==0.4.6
git+https://FETKlOkAn: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@github.com/dazehead/Coinbase_Trader.git@db2bd1c55e431f2e7e2c1787b472bda8b5f89825#egg=coinbase_trader
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cupy-cuda12x==13.3.0
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
fastrlock==0.8.2
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.53.1
idna==3.8
imageio==2.35.1
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
llist==0.7.1
llvmlite==0.39.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numba==0.56.4
numpy==1.23.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
plotly==5.24.0
prompt_toolkit==3.0.47
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
ta-lib==0.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vectorbt==0.26.2
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
